BE/requirements.txt
# AutoSchemaKG 백엔드 서버 의존성
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0

# Neo4j 관련
neo4j>=5.8.0
graphdatascience>=1.9.0

# AI/ML 관련
openai>=1.0.0
sentence-transformers>=2.7.0
transformers>=4.42.0
torch>=2.0.0
numpy>=1.26.0
pandas>=2.0.0

# 데이터 처리
tqdm>=4.65.0
jsonschema>=4.17.3
networkx>=3.1
nltk>=3.8.0

# 유틸리티
tenacity>=9.1.2
einops>=0.7.0
accelerate>=1.7.0
json_repair>=0.7.0

# # Azure 관련 (선택사항)
# azure-ai-projects
# azure-cli>=2.50.0
# azure-ai-inference>=1.0.0
# azure-identity>=1.15.0
